gerald/pyproject.toml
[project]
name = "gerald"
version = "0.1.0"
description = "Gerald – an AI-assisted X (Twitter) opportunity discovery and outreach copilot."
authors = [{ name = "Jason Mellet" }]
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy>=2.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "typer[all]>=0.12",
  "httpx>=0.27",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "anthropic>=0.39",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

